@@ -5,8 +5,8 @@
   "main": "karma.conf.js",
   "scripts": {
     "start:app": "ng serve --host=0.0.0.0 --disable-host-check --proxy-config proxy.config.js",
-    "build:prod": "ng build ---env=prod",
-    "build:test": "ng build ---env=test",
+    "build:prod": "ng build ---configuration=production",
+    "build:test": "ng build ---configuration=test",
     "test": "ng test",
     "lint": "ng lint",
     "e2e": "ng e2e"
@@ -15,45 +15,50 @@
   "license": "MIT",
   "private": true,
   "dependencies": {
-    "@angular/animations": "^5.0.5",
-    "@angular/cdk": "^2.0.0-beta.8",
-    "@angular/common": "^5.0.5",
-    "@angular/compiler": "^5.0.5",
-    "@angular/core": "^5.0.5",
-    "@angular/flex-layout": "2.0.0-beta.12",
-    "@angular/forms": "^5.0.5",
-    "@angular/http": "^5.0.5",
-    "@angular/material": "^2.0.0-beta.8",
-    "@angular/platform-browser": "^5.0.5",
-    "@angular/platform-browser-dynamic": "^5.0.5",
-    "@angular/platform-server": "^5.0.5",
-    "@angular/router": "^5.0.5",
+    "@angular/animations": "^7.2.3",
+    "@angular/cdk": "^7.3.0",
+    "@angular/common": "^7.2.3",
+    "@angular/compiler": "^7.2.3",
+    "@angular/core": "^7.2.3",
+    "@angular/flex-layout": "^7.0.0-beta.22",
+    "@angular/forms": "^7.2.3",
+    "@angular/http": "^7.2.3",
+    "@angular/material": "^7.3.0",
+    "@angular/platform-browser": "^7.2.3",
+    "@angular/platform-browser-dynamic": "^7.2.3",
+    "@angular/platform-server": "^7.2.3",
+    "@angular/router": "^7.2.3",
     "@ngx-translate/core": "^8.0.0",
     "@ngx-translate/http-loader": "^2.0.0",
     "@types/file-saver": "^1.3.0",
-    "angular2-markdown": "^2.2.0",
+    "angular2-markdown": "^2.2.3",
     "bootstrap": "^3.3.7",
     "bootstrap-accessibility-plugin": "^1.0.6",
+    "classlist.js": "^1.1.20150312",
     "core-js": "^2.4.1",
     "enhanced-resolve": "^3.3.0",
     "file-saver": "^1.3.8",
     "jquery": "^3.3.1",
     "ng2-file-upload": "^1.2.1",
     "ngx-popover": "0.0.16",
-    "rxjs": "^5.5.4",
+    "rxjs": "^6.4.0",
+    "rxjs-compat": "^6.0.0-rc.0",
     "sweetalert2": "^7.26.29",
+    "tslib": "^1.9.0",
     "typings": "^2.1.1",
-    "zone.js": "^0.8.14"
+    "web-animations-js": "^2.3.1",
+    "zone.js": "^0.8.29"
   },
   "devDependencies": {
-    "@angular/cli": "^1.2.6",
-    "@angular/compiler-cli": "^5.0.5",
-    "@angular/language-service": "^4.0.0",
+    "@angular-devkit/build-angular": "~0.13.0",
+    "@angular/cli": "^7.3.0",
+    "@angular/compiler-cli": "^7.2.3",
+    "@angular/language-service": "^7.2.3",
     "@types/jasmine": "~2.5.53",
     "@types/jasminewd2": "~2.0.2",
-    "@types/jquery": "^3.3.0",
+    "@types/jquery": "^3.3.6",
     "@types/node": "~6.0.60",
-    "codelyzer": "~3.0.1",
+    "codelyzer": "^4.5.0",
     "jasmine-core": "~2.6.2",
     "jasmine-spec-reporter": "~4.1.0",
     "karma": "~1.7.0",
@@ -65,6 +70,6 @@
     "protractor": "~5.1.2",
     "ts-node": "~3.0.4",
     "tslint": "~5.3.2",
-    "typescript": "2.4.2"
+    "typescript": "3.2.4"
   }
 }
